Migrated to use Long instead of Integer, increasing to 10 trillion limit.

+++ src/main/java/io/felux/piggybanks/api/PiggyBank.java
@@ -18,7 +18,7 @@ public class PiggyBank {
 
     private static PiggyBanks plugin = (PiggyBanks) JavaPlugin.getProvidingPlugin(PiggyBanks.class);
 
-    public static ItemStack create(String name, UUID uuid, int amount) {
+    public static ItemStack create(String name, UUID uuid, long amount) {
         String altName = name;
         if (name.equalsIgnoreCase("CONSOLE")) {
             name = Lang.CONSOLE.asString();
@@ -41,7 +41,7 @@ public class PiggyBank {
         if (nbtItem != null) {
             nbtItem.setString("created-by-name", altName);
             nbtItem.setString("created-by-uuid", uuid.toString());
-            nbtItem.setInt("balance", amount);
+            nbtItem.setLong("balance", amount);
         }
 
         String texture = plugin.getConfig().getString("skull", "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvMTk4ZGY0MmY0NzdmMjEzZmY1ZTlkN2ZhNWE0Y2M0YTY5ZjIwZDljZWYyYjkwYzRhZTRmMjliZDE3Mjg3YjUifX19");
